// src/api-services/models/create-customer-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 规范化接口演示
 * 让 .NET 开发更简单，更通用，更流行。
 *
 * OpenAPI spec version: 1.0.0
 * Contact: monksoul@outlook.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateCustomerDto
 */
export interface CreateCustomerDto {
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    companyName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    nit?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    ci?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    phone?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    address?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    city?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    state?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    country?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCustomerDto
     */
    description?: string | null;
}